Move Constraint to torch.compiler namespace
